+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 9.23.0 -
 ### Added
 - Added Altar of Fangs, Venomous Abyss and Tidebound Grotto instance loot
+- Added Zul'jarra's Forces and Captain Tokka to Factions
+- Added Prey and Delve season journeys to Factions
 - Added Covenant renowns to Factions
 - Added quest sources to some DF tabards